feature-chart updates
