+++ dev/joymeter-android-client/AndroidClient/app/src/main/res/layout/activity_chart.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent">
 
     <com.github.mikephil.charting.charts.LineChart
-        android:id="@+id/chart1"
+        android:id="@+id/chart"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_above="@+id/suggestButton" />
